refactor: File change tracking and path handling
refactor: File change tracking and path handling

Moved file creation, modification, and deletion tracking from ClientHandler to FileHandler using thread-safe sets. Added utility methods for converting between network and local path strings to improve cross-platform support. Improved temporary file handling and ensured proper cleanup. Simplified ClientHandler by removing file tracking logic.

+++ src/main/java/com/rivelbop/dossio/io/FileHandler.java
@@ -17,9 +17,12 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Stream;
 import javax.annotation.CheckForNull;
 import org.eclipse.jgit.diff.EditList;
@@ -35,7 +38,7 @@ public final class FileHandler {
   private final Path projectDirectoryPath;
 
   private final Path tempDirectoryPath;
-  private final HashMap<Path, Path> cachedTempPaths = new HashMap<>();
+  private final ConcurrentHashMap<Path, Path> cachedTempPaths = new ConcurrentHashMap<>();
 
   private final FileWatcher fileWatcher;
   private final FileFilter fileFilter;
@@ -43,6 +46,10 @@ public final class FileHandler {
   private final EditInterpreter editInterpreter = new EditInterpreter();
 
   private final ClientHandler clientHandler = Main.NETWORK.getClientHandler();
+  private final Set<String> filesMarkedForCreation = Collections.synchronizedSet(new HashSet<>());
+  private final Set<String> filesMarkedForModification =
+      Collections.synchronizedSet(new HashSet<>());
+  private final Set<String> filesMarkedForDeletion = Collections.synchronizedSet(new HashSet<>());
 
   /**
    * Creates a file handler (initializes the watcher and filter).
@@ -134,6 +141,30 @@ public final class FileHandler {
   }
 
   /**
+   * Converts a path to its Unix string representation to send over the network.
+   *
+   * <p>This is used to ensure cross-platform support, this string can then be converted to the
+   * platform it is used for.
+   *
+   * @param path The path to represent as Unix string path.
+   * @return The path's Unix-based network string representation.
+   */
+  public static String pathToNetworkString(Path path) {
+    return path.toString().replace(File.separatorChar, '/');
+  }
+
+  /**
+   * Converts a network path string to a local one by replacing the Unix-based file separator with
+   * the system's file separator.
+   *
+   * @param pathStr The network path string to convert (should be Unix-based).
+   * @return The local string representation of the network path string.
+   */
+  public static String networkPathStringToLocalPathString(String pathStr) {
+    return pathStr.replace('/', File.separatorChar);
+  }
+
+  /**
    * Called when file creation is detected.
    *
    * @param absoluteFilePath The absolute path of the created file.
@@ -159,8 +190,8 @@ public final class FileHandler {
     }
 
     // Ensure no creation packet is sent if you received a packet to create this file
-    String relativePathStr = relativeFilePath.toString();
-    if (clientHandler.getFilesMarkedForCreation().remove(relativePathStr)) {
+    String relativePathStr = pathToNetworkString(relativeFilePath);
+    if (filesMarkedForCreation.remove(relativePathStr)) {
       return;
     }
 
@@ -216,12 +247,14 @@ public final class FileHandler {
    * @throws RuntimeException If error occurs when copying modified data to temporary file.
    */
   public void onModify(Path absoluteFilePath) {
-    // This avoids potential issues when a file is deleted before onModify is called
-    if (!Files.exists(absoluteFilePath)) {
+    Path relativeFilePath = projectDirectoryPath.relativize(absoluteFilePath);
+    String fileName = pathToNetworkString(relativeFilePath);
+
+    // This avoids checking the changes in a modification from the network, while also avoiding
+    // potential issues when a file is deleted before onModify is called
+    if (filesMarkedForModification.remove(fileName) || !Files.exists(absoluteFilePath)) {
       return;
     }
-
-    Path relativeFilePath = projectDirectoryPath.relativize(absoluteFilePath);
 
     // Check if file is ignored
     if (fileFilter.isIgnored(relativeFilePath, absoluteFilePath)) {
@@ -255,7 +288,6 @@ public final class FileHandler {
     }
 
     // Convert the changes into packets
-    String fileName = projectDirectoryPath.relativize(absoluteFilePath).toString();
     List<EditPacket> editPackets = EditSerializer.toEditPackets(fileName, newLines, editList);
 
     // Begin sending edit packets to the server
@@ -286,7 +318,6 @@ public final class FileHandler {
    * Called when file deletion is detected.
    *
    * @param absoluteFilePath The absolute path of the deleted file.
-   * @throws RuntimeException If IO error occurs when deleting the temporary text file.
    */
   public void onDelete(Path absoluteFilePath) {
     Path relativeFilePath = projectDirectoryPath.relativize(absoluteFilePath);
@@ -298,21 +329,13 @@ public final class FileHandler {
 
     Log.info(LOG_TAG, "DELETED: " + absoluteFilePath);
 
-    // Get the temporary text file (if it exists) and delete it
-    Path tempFile = getTempPath(absoluteFilePath);
-    if (tempFile != null) {
-      try {
-        Files.delete(tempFile);
-        cachedTempPaths.remove(absoluteFilePath);
-      } catch (IOException e) {
-        Log.error(LOG_TAG, "Failed to delete temporary text file!", e);
-        throw new RuntimeException(e); // This may cause issues if the same temp file is recreated
-      }
-    }
+    // Get the temporary text file (if it exists)
+    // Automatically deletes the file (if the file at absoluteFilePath doesn't exist)
+    getTempPath(absoluteFilePath);
 
     // Send delete file packet to server
-    String relativePathStr = relativeFilePath.toString();
-    if (!clientHandler.getFilesMarkedForDeletion().remove(relativePathStr)
+    String relativePathStr = pathToNetworkString(relativeFilePath);
+    if (!filesMarkedForDeletion.remove(relativePathStr)
         && Files.exists(absoluteFilePath.getParent())) {
       DeleteFilePacket deletePacket = new DeleteFilePacket();
       deletePacket.fileName = relativePathStr;
@@ -332,7 +355,8 @@ public final class FileHandler {
     } else if (o instanceof EditPacket p) {
       editInterpreter.insert(p);
     } else if (o instanceof EndEditPacket p) {
-      Path absFilePath = projectDirectoryPath.resolve(p.fileName);
+      Path absFilePath =
+          projectDirectoryPath.resolve(networkPathStringToLocalPathString(p.fileName));
       List<String> lines;
 
       // Read file lines and apply the edits
@@ -343,6 +367,8 @@ public final class FileHandler {
         return;
       }
       editInterpreter.apply(editInterpreter.end(p), lines);
+
+      filesMarkedForModification.add(p.fileName);
 
       // Write the updated lines to both the temporary and actual files
       try {
@@ -358,14 +384,18 @@ public final class FileHandler {
   /**
    * Creates a specified relative project file.
    *
+   * <p>NOTE: This is supposed to be used to create a file from the network.
+   *
    * @param fileName The name of the file to create (relative to project directory).
    */
   public void createFile(String fileName) {
-    Path absFilePath = projectDirectoryPath.resolve(fileName);
+    Path absFilePath = projectDirectoryPath.resolve(networkPathStringToLocalPathString(fileName));
+    if (Files.exists(absFilePath)) {
+      return;
+    }
+
+    filesMarkedForCreation.add(fileName);
     try {
-      if (Files.exists(absFilePath)) {
-        return;
-      }
       Files.createDirectories(absFilePath.getParent()); // Ensure parent directories exist
       Files.createFile(absFilePath);
     } catch (IOException e) {
@@ -376,12 +406,22 @@ public final class FileHandler {
   /**
    * Deletes a specified relative project file.
    *
+   * <p>NOTE: This is supposed to be used to delete a file from the network.
+   *
    * @param fileName The name of the file to delete (relative to project directory).
    */
   public void deleteFile(String fileName) {
-    Path absFilePath = projectDirectoryPath.resolve(fileName);
+    Path absFilePath = projectDirectoryPath.resolve(networkPathStringToLocalPathString(fileName));
+    if (!Files.exists(absFilePath)) {
+      return;
+    }
+
+    filesMarkedForDeletion.add(fileName);
     try {
+      // Delete if directory
       FileUtils.delete(absFilePath.toFile(), FileUtils.RECURSIVE | FileUtils.SKIP_MISSING);
+
+      // Delete if file (in the case that FileUtils didn't delete it)
       Files.deleteIfExists(absFilePath);
     } catch (IOException e) {
       Log.error(LOG_TAG, "Failed to delete file!", e);
@@ -405,20 +445,40 @@ public final class FileHandler {
   private Path getTempPath(Path absolutePath) {
     // Check if a temporary path was already cached
     Path tempPath = cachedTempPaths.get(absolutePath);
+
+    // Ensure the file exists and is a text file
+    if (!Files.exists(absolutePath) || !isTextFile(absolutePath)) {
+      // If it used to, remove its temp path
+      if (tempPath != null) {
+        cachedTempPaths.remove(tempPath);
+        try {
+          Files.delete(tempPath);
+        } catch (IOException e) {
+          Log.error(LOG_TAG, "Failed to delete temporary text file!", e);
+          return null;
+        }
+      }
+      return null;
+    }
+
+    // If the file is valid and the path is already cached
     if (tempPath != null) {
       return tempPath;
     }
 
-    // Ensure the file exists and is a text file
-    if (!Files.exists(absolutePath) || !isTextFile(absolutePath)) {
-      return null;
-    }
-
     // Get the name of the path relative to the project directory path
     String relativePathName = projectDirectoryPath.relativize(absolutePath).toString();
+    // Remove any "special" characters from the file name
+    String safePathName =
+        relativePathName.replace(File.separatorChar, '_').replaceAll("[^a-zA-Z0-9._-]", "_");
+    // Ensure file name under 250 chars (Windows limit)
+    // Go under 200 here to add the hash
+    if (safePathName.length() > 200) {
+      safePathName = safePathName.substring(0, 200);
+    }
 
     // Create the temporary path if it wasn't already cached
-    String tempPathName = relativePathName.replaceAll("/", "_").concat(".tmp");
+    String tempPathName = safePathName + "_" + relativePathName.hashCode() + ".tmp";
     try {
       tempPath = Files.copy(absolutePath, tempDirectoryPath.resolve(tempPathName), COPY_ATTRIBUTES);
     } catch (IOException e) {
@@ -443,6 +503,7 @@ public final class FileHandler {
   private void registerDirectory(Path absDirectoryPath, boolean alertNetworkOfNestedFiles) {
     try (Stream<Path> stream = Files.walk(absDirectoryPath)) {
       stream
+          .filter(path -> !Files.isSymbolicLink(path))
           .filter(path -> !path.startsWith(tempDirectoryPath))
           .filter(path -> !fileFilter.isIgnored(projectDirectoryPath.relativize(path), path))
           .forEach(
